refactor: Header components structure
